fix: resolve all MC 1.21.11 compilation errors
- SpearDamage: add missing Vec3 import
- TracerRenderer: add missing Vec3 import
- MaceDamage: fix double/float type mismatch for fallDistance
- Reach: fix boolean logic on onlyWhenAttacking check
- Clutch: fix float/int type mismatch for getHealth()
- BlockESP/MobESP: stub render methods (3D projection TODO)

All compilation errors should now be resolved.

+++ src/main/java/com/sigterm/module/combat/Clutch.java
@@ -24,7 +24,7 @@ public class Clutch extends Module {
     public void onTick() {
         if (mc().player == null) return;
 
-        int health = mc().player.getHealth();
+        float health = mc().player.getHealth();
         int threshold = (int) healthThreshold.value;
         boolean shouldClutch = health <= threshold;
